requirements.txt
torch
torchvision
diffusers==0.34.0
transformers
accelerate
peft==0.15.0
datasets
wandb
Pillow
pillow-heif
numpy
pandas>=2.0.0
tqdm
tensorboard>=2.13.0
safetensors>=0.3.0
gradio
requests
spaces